Build for 32-bit too
git-svn-id: https://svn.ices.utexas.edu/repos/iss/projects/GaloisCpp/trunk@9330 522c16a8-e32f-0410-b4cc-8071c0b41197

+++ include/Galois/Bag.h
@@ -242,7 +242,7 @@ public:
     }
 
     // ------- random access iterator concepts -------
-    typedef long difference_type;
+    typedef ptrdiff_t difference_type;
 
     iterator& operator+=(difference_type n) {
       while (n > 0) {
